+++ bin/actions/0220022-1.sh
@@ -3,7 +3,7 @@
 #: ◢ + Esc - Bring System Monitor to front, launch if not yet running
 
 if pidof -s gnome-system-monitor > /dev/null; then
-    systemMonitorPid=$(pidof nemo)
+    systemMonitorPid=$(pidof gnome-system-monitor)
     windowTitle=$( wmctrl -l -p | grep "$systemMonitorPid" | cut -d " " -f 8-)
 	wmctrl -a "$windowTitle" &> /dev/null
 else