+++ bin/actions/0220022-23.sh
@@ -7,7 +7,6 @@ focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
 
 path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
-
 "$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
 
 public_ip=$(curl -s ident.me)