+++ bin/actions/0220022-33.sh
@@ -9,7 +9,6 @@ focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
 
 path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
-
 "$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
 
 if [[ "$focusApplication" =~ "claws-mail" ]]; then