+++ bin/actions/0220022-49.sh
@@ -6,7 +6,6 @@ focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
 
 path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
-
 "$path_self"/utils/remove_unwanted_output.sh "$focusApplication"
 
 if pidof -s nemo > /dev/null; then