+++ CHANGELOG.md
@@ -5,6 +5,7 @@ V0.4.3 - Not tagged yet
 -----------------------
 * Improve actions' compatibility: Handle variations of Chromium's application identifier
 * Add xpad to applications removing undesired output of characters from chording combo
+* Add action: ◢ + Y - Bring Yana to front, launch if not yet running
 
 V0.4.2 - 2021/01/29
 -------------------